+++ app/Http/Controllers/FoodScheduleController.php
@@ -4,23 +4,10 @@ namespace App\Http\Controllers;
 
 use App\FoodSchedule;
 use Illuminate\Http\Request;
+use Illuminate\Auth\Access\Gate;
 
 class FoodScheduleController extends Controller
 {
-    public function create(Request $request)
-    {
-        $createFoodSchedule = new FoodSchedule;
-        $createFoodSchedule->schedule_date = $request->schedule_date;
-        $createFoodSchedule->id_user = $request->id_user;
-        $createFoodSchedule->id_food = $request->id_food;
-        $createFoodSchedule->save();
-
-        return response([
-            'status' => 'OK',
-            'message' => 'Food Schedule telah dibuat',
-            'data' => $createFoodSchedule
-        ], 200);
-    }
 
     public function food_schedules()
     {
@@ -28,22 +15,47 @@ class FoodScheduleController extends Controller
         return response($food_schedules, 200);
     }
 
+    public function create(Request $request)
+    {
+        if (Gate::allows('admin-only'))
+        {
+            // Hanya User dengan role admin yang dapat mengakses ini
+            $createFoodSchedule = new FoodSchedule;
+            $createFoodSchedule->schedule_date = $request->schedule_date;
+            $createFoodSchedule->id_user = $request->id_user;
+            $createFoodSchedule->id_food = $request->id_food;
+            $createFoodSchedule->save();
+            return response()->json(["message" => "Sukses membuat produk", "data" => $createFoodSchedule], 201);
+        }
+        return response()->json(["message" => "Anda tidak memiliki akses"], 403);
+
+    }
+
     public function update(Request $request, $id)
     {
-        $updateFoodSchedule = FoodSchedule::find($id);
 
-        $updateFoodSchedule->update($request->all());
-        return response($updateFoodSchedule, 200);
+        if (Gate::allows('admin-only'))
+        {
+            // Hanya User dengan role admin yang dapat mengakses ini
+            $updateFoodSchedule = FoodSchedule::find($id);
+
+            $updateFoodSchedule->update($request->all());
+            return response()->json(["message" => "Produk berhasil diubah", "data" => $updateFoodSchedule], 201);
+        }
+        return response()->json(["message" => "Anda tidak memiliki akses"], 403);
+
     }
 
     public function delete($id)
     {
-        $deleteFoodSchedule = FoodSchedule::find($id);
-        $deleteFoodSchedule->delete();
 
-        return response([
-            'status' => 'OK',
-            'message' => 'Food Schedule telah dihapus',
-        ], 202);
+        if (Gate::allows('admin-only')) {
+            // Hanya User dengan role admin yang dapat mengakses ini    
+            $deleteFoodSchedule = FoodSchedule::find($id);
+            $deleteFoodSchedule->delete();
+            return response()->json(["message" => "Produk telah dihapus"], 201);
+        }
+        return response()->json(["message" => "Anda tidak memiliki akses"], 403);
+
     }
 }
